Descr identity, EffectInfo bitstring overhaul, and cache-key widening
Process-global GcCache becomes the single source of truth for size /
field / array / arraylen / call / interiorfield descrs.  Every mint
site routes through it; the descr_registry facade publishes external
mints (legacy paths predating the keyed cache).  Per-call caches now
key on identity instead of structural equality: ARRAY len descrs by
ARRAY identity, struct descr groups by STRUCT identity, call descrs
by full call signature.  Type-id-less mints (legacy make_array_descr,
bh_size_spec_from_callcontrol) mint a fresh Arc per call instead of
collapsing onto a shared bucket, matching PyPy's per-identity
_cache_arraylen / _cache_size semantics.  Callcontrol struct identity
stamps path_hash(owner) as the type_id surrogate.

__majit_type_id switches from TypeId hashing to
path_hash(module_path!() + "::" + StructName) so analyzer-time and
runtime-time mint sites can produce matching LLType::Struct(_) keys.

BhDescr serialization layer widens cache-key carriers from u32 to
u64: BhSizeSpec.type_id, BhDescr::{Size,Array}.type_id,
AssemblerDescrKey, DispatchArrayDescrKey, SIMPLE_DESCR_GROUP_CACHE.
bh_size_spec_from_callcontrol drops the `as u32` truncation and
passes the full path_hash(owner) u64.  Downstream backends that
allocate via gc.alloc_*_typed keep the u32 GC tid signature; bridge
sites narrow `as u32` with a structural-fix marker.

CallControl::arraydescrof threads path_hash(array_type_id) as u32
into the SimpleArrayDescr tid slot so distinct ARRAYs no longer
collapse onto tid=0.  set_all_interiorfielddescrs is promoted to the
ArrayDescr trait with a default no-op so callers holding
Arc<dyn Descr> can dispatch through as_array_descr().

EffectInfo gains a structural-identity DescrGroup and per-extraeffect
shapes: Plain, PlainCannotRaise, CannotRaiseNoHeap, MayForce, and
RandomEffects each carry distinct ei_index bitstrings.  The raw-set
type widens from Vec<u32> to Vec<DescrRef>, with subtract operations
keyed on Arc pointer identity instead of descr.index().
Analyzer-side WriteAnalysis populates _readonly/write_descrs_fields /
_arrays / _interiorfields via cc.fielddescrof /
cc.interiorfielddescrof.  compute_bitstrings allocates ei_index per
descr-set partition, writes descr.set_ei_index(), and publishes a
fallback descr.index() side table for legacy heap.rs readers.

heap.rs field_effect_index / array_effect_index read
descr.get_ei_index() first then the side table.
ParentBackedFieldDescr delegates ei_index to its inner
SimpleFieldDescr so wrapper clones see the same storage.
finish_setup_descrs assigns sequential descr_index across all
categories and seals legacy bitstring writes; post-setup mints with
non-trivial raw sets panic.  JitCellToken.target_tokens routes
has_compiled_loop through the JCT-side TargetToken descrs.

Flatten tests retire the MOST_GENERAL collapse assumption: Plain /
PlainCannotRaise / MayForce now carry distinct EffectInfo identities,
build_load_global helpers no longer assert equality with binary-op
helpers, and the round-trip / check_can_raise tests cover all
extraeffect shapes.

+++ majit/majit-backend-dynasm/src/runner.rs
@@ -2166,7 +2166,10 @@ impl Backend for DynasmBackend {
         // allocation requires a real GC type id; tid=0 means the descr
         // never went through `gc.py:548 set_type_id` and the GC tracer
         // would lack the per-item visit shape.
-        let type_id = arraydescr.get_type_id();
+        // PRE-EXISTING-ADAPTATION: `BhDescr.get_type_id()` returns the
+        // u64 `path_hash` cache key, but `dynasm_alloc_*` expects the
+        // u32 GC tid.  Truncate `as u32` until gc_cache routing.
+        let type_id = arraydescr.get_type_id() as u32;
         assert!(
             type_id != 0,
             "bh_new_array requires ArrayDescr.tid (descr.py:340) — got 0"
